+++ lib/cusparse/CUSPARSE.jl
@@ -16,12 +16,19 @@ include("error.jl")
 include("libcusparse.jl")
 include("libcusparse_old.jl")
 
-# low-level wrappers
 include("array.jl")
 include("util.jl")
 include("types.jl")
+
+# low-level wrappers
 include("helpers.jl")
-include("wrappers.jl")
+include("management.jl")
+include("level1.jl")
+include("level2.jl")
+include("level3.jl")
+include("preconditioners.jl")
+include("conversions.jl")
+include("generic.jl")
 
 # high-level integrations
 include("interfaces.jl")
