+++ lib/cusparse/preconditioners.jl
@@ -1,0 +1,266 @@
+# routines that implement different preconditioners
+
+export ic02!, ic02, ilu02!, ilu02
+
+"""
+    ic02!(A::CuSparseMatrix, index::SparseChar)
+
+Incomplete Cholesky factorization with no pivoting.
+Preserves the sparse layout of matrix `A`.
+"""
+ic02!(A::CuSparseMatrix, index::SparseChar)
+for (bname,aname,sname,elty) in ((:cusparseScsric02_bufferSize, :cusparseScsric02_analysis, :cusparseScsric02, :Float32),
+                                 (:cusparseDcsric02_bufferSize, :cusparseDcsric02_analysis, :cusparseDcsric02, :Float64),
+                                 (:cusparseCcsric02_bufferSize, :cusparseCcsric02_analysis, :cusparseCcsric02, :ComplexF32),
+                                 (:cusparseZcsric02_bufferSize, :cusparseZcsric02_analysis, :cusparseZcsric02, :ComplexF64))
+    @eval begin
+        function ic02!(A::CuSparseMatrixCSR{$elty},
+                       index::SparseChar)
+            desc = CuMatrixDescriptor(CUSPARSE_MATRIX_TYPE_GENERAL, CUSPARSE_FILL_MODE_LOWER, CUSPARSE_DIAG_TYPE_NON_UNIT, index)
+            m,n = A.dims
+            if m != n
+                throw(DimensionMismatch("A must be square, but has dimensions ($m,$n)!"))
+            end
+            info = csric02Info_t[0]
+            cusparseCreateCsric02Info(info)
+            @workspace size=@argout(
+                    $bname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, info[1],
+                           out(Ref{Cint}(1)))
+                )[] buffer->begin
+                    $aname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                    posit = Ref{Cint}(1)
+                    cusparseXcsric02_zeroPivot(handle(), info[1], posit)
+                    if posit[] >= 0
+                        error("Structural/numerical zero in A at ($(posit[]),$(posit[])))")
+                    end
+                    $sname(handle(), m, A.nnz,
+                           desc, A.nzVal, A.rowPtr, A.colVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                end
+            cusparseDestroyCsric02Info(info[1])
+            A
+        end
+    end
+end
+
+# cscic02
+for (bname,aname,sname,elty) in ((:cusparseScsric02_bufferSize, :cusparseScsric02_analysis, :cusparseScsric02, :Float32),
+                                 (:cusparseDcsric02_bufferSize, :cusparseDcsric02_analysis, :cusparseDcsric02, :Float64),
+                                 (:cusparseCcsric02_bufferSize, :cusparseCcsric02_analysis, :cusparseCcsric02, :ComplexF32),
+                                 (:cusparseZcsric02_bufferSize, :cusparseZcsric02_analysis, :cusparseZcsric02, :ComplexF64))
+    @eval begin
+        function ic02!(A::CuSparseMatrixCSC{$elty},
+                       index::SparseChar)
+            desc = CuMatrixDescriptor(CUSPARSE_MATRIX_TYPE_GENERAL, CUSPARSE_FILL_MODE_LOWER, CUSPARSE_DIAG_TYPE_NON_UNIT, index)
+            m,n = A.dims
+            if m != n
+                throw(DimensionMismatch("A must be square, but has dimensions ($m,$n)!"))
+            end
+            info = csric02Info_t[0]
+            cusparseCreateCsric02Info(info)
+            @workspace size=@argout(
+                    $bname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.colPtr, A.rowVal, info[1],
+                           out(Ref{Cint}(1)))
+                )[] buffer->begin
+                    $aname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.colPtr, A.rowVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                    posit = Ref{Cint}(1)
+                    cusparseXcsric02_zeroPivot(handle(), info[1], posit)
+                    if posit[] >= 0
+                        error("Structural/numerical zero in A at ($(posit[]),$(posit[])))")
+                    end
+                    $sname(handle(), m, A.nnz,
+                           desc, A.nzVal, A.colPtr, A.rowVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                end
+            cusparseDestroyCsric02Info(info[1])
+            A
+        end
+    end
+end
+
+"""
+    ilu02!(A::CuSparseMatrix, index::SparseChar)
+
+Incomplete LU factorization with no pivoting.
+Preserves the sparse layout of matrix `A`.
+"""
+ilu02!(A::CuSparseMatrix, index::SparseChar)
+for (bname,aname,sname,elty) in ((:cusparseScsrilu02_bufferSize, :cusparseScsrilu02_analysis, :cusparseScsrilu02, :Float32),
+                                 (:cusparseDcsrilu02_bufferSize, :cusparseDcsrilu02_analysis, :cusparseDcsrilu02, :Float64),
+                                 (:cusparseCcsrilu02_bufferSize, :cusparseCcsrilu02_analysis, :cusparseCcsrilu02, :ComplexF32),
+                                 (:cusparseZcsrilu02_bufferSize, :cusparseZcsrilu02_analysis, :cusparseZcsrilu02, :ComplexF64))
+    @eval begin
+        function ilu02!(A::CuSparseMatrixCSR{$elty},
+                        index::SparseChar)
+            desc = CuMatrixDescriptor(CUSPARSE_MATRIX_TYPE_GENERAL, CUSPARSE_FILL_MODE_LOWER, CUSPARSE_DIAG_TYPE_NON_UNIT, index)
+            m,n = A.dims
+            if m != n
+                throw(DimensionMismatch("A must be square, but has dimensions ($m,$n)!"))
+            end
+            info = csrilu02Info_t[0]
+            cusparseCreateCsrilu02Info(info)
+            @workspace size=@argout(
+                    $bname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, info[1],
+                           out(Ref{Cint}(1)))
+                )[] buffer->begin
+                    $aname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                    posit = Ref{Cint}(1)
+                    cusparseXcsrilu02_zeroPivot(handle(), info[1], posit)
+                    if posit[] >= 0
+                        error("Structural zero in A at ($(posit[]),$(posit[])))")
+                    end
+                    $sname(handle(), m, A.nnz,
+                           desc, A.nzVal, A.rowPtr, A.colVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                end
+            cusparseDestroyCsrilu02Info(info[1])
+            A
+        end
+    end
+end
+
+# cscilu02
+for (bname,aname,sname,elty) in ((:cusparseScsrilu02_bufferSize, :cusparseScsrilu02_analysis, :cusparseScsrilu02, :Float32),
+                                 (:cusparseDcsrilu02_bufferSize, :cusparseDcsrilu02_analysis, :cusparseDcsrilu02, :Float64),
+                                 (:cusparseCcsrilu02_bufferSize, :cusparseCcsrilu02_analysis, :cusparseCcsrilu02, :ComplexF32),
+                                 (:cusparseZcsrilu02_bufferSize, :cusparseZcsrilu02_analysis, :cusparseZcsrilu02, :ComplexF64))
+    @eval begin
+        function ilu02!(A::CuSparseMatrixCSC{$elty},
+                        index::SparseChar)
+            desc = CuMatrixDescriptor(CUSPARSE_MATRIX_TYPE_GENERAL, CUSPARSE_FILL_MODE_LOWER, CUSPARSE_DIAG_TYPE_NON_UNIT, index)
+            m,n = A.dims
+            if m != n
+                throw(DimensionMismatch("A must be square, but has dimensions ($m,$n)!"))
+            end
+            info = csrilu02Info_t[0]
+            cusparseCreateCsrilu02Info(info)
+            @workspace size=@argout(
+                    $bname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.colPtr, A.rowVal, info[1],
+                           out(Ref{Cint}(1)))
+                )[] buffer->begin
+                    $aname(handle(), m, A.nnz, desc,
+                           A.nzVal, A.colPtr, A.rowVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                    posit = Ref{Cint}(1)
+                    cusparseXcsrilu02_zeroPivot(handle(), info[1], posit)
+                    if posit[] >= 0
+                        error("Structural zero in A at ($(posit[]),$(posit[])))")
+                    end
+                    $sname(handle(), m, A.nnz,
+                           desc, A.nzVal, A.colPtr, A.rowVal, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                end
+            cusparseDestroyCsrilu02Info(info[1])
+            A
+        end
+    end
+end
+
+# bsric02
+for (bname,aname,sname,elty) in ((:cusparseSbsric02_bufferSize, :cusparseSbsric02_analysis, :cusparseSbsric02, :Float32),
+                                 (:cusparseDbsric02_bufferSize, :cusparseDbsric02_analysis, :cusparseDbsric02, :Float64),
+                                 (:cusparseCbsric02_bufferSize, :cusparseCbsric02_analysis, :cusparseCbsric02, :ComplexF32),
+                                 (:cusparseZbsric02_bufferSize, :cusparseZbsric02_analysis, :cusparseZbsric02, :ComplexF64))
+    @eval begin
+        function ic02!(A::CuSparseMatrixBSR{$elty},
+                       index::SparseChar)
+            desc = CuMatrixDescriptor(CUSPARSE_MATRIX_TYPE_GENERAL, CUSPARSE_FILL_MODE_UPPER, CUSPARSE_DIAG_TYPE_NON_UNIT, index)
+            m,n = A.dims
+            if m != n
+                throw(DimensionMismatch("A must be square, but has dimensions ($m,$n)!"))
+            end
+            mb = div(m,A.blockDim)
+            info = bsric02Info_t[0]
+            cusparseCreateBsric02Info(info)
+            @workspace size=@argout(
+                    $bname(handle(), A.dir, mb, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, A.blockDim, info[1],
+                           out(Ref{Cint}(1)))
+                )[] buffer->begin
+                    $aname(handle(), A.dir, mb, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, A.blockDim, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                    posit = Ref{Cint}(1)
+                    cusparseXbsric02_zeroPivot(handle(), info[1], posit)
+                    if posit[] >= 0
+                        error("Structural/numerical zero in A at ($(posit[]),$(posit[])))")
+                    end
+                    $sname(handle(), A.dir, mb, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, A.blockDim, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                end
+            cusparseDestroyBsric02Info(info[1])
+            A
+        end
+    end
+end
+
+# bsrilu02
+for (bname,aname,sname,elty) in ((:cusparseSbsrilu02_bufferSize, :cusparseSbsrilu02_analysis, :cusparseSbsrilu02, :Float32),
+                                 (:cusparseDbsrilu02_bufferSize, :cusparseDbsrilu02_analysis, :cusparseDbsrilu02, :Float64),
+                                 (:cusparseCbsrilu02_bufferSize, :cusparseCbsrilu02_analysis, :cusparseCbsrilu02, :ComplexF32),
+                                 (:cusparseZbsrilu02_bufferSize, :cusparseZbsrilu02_analysis, :cusparseZbsrilu02, :ComplexF64))
+    @eval begin
+        function ilu02!(A::CuSparseMatrixBSR{$elty},
+                        index::SparseChar)
+            desc = CuMatrixDescriptor(CUSPARSE_MATRIX_TYPE_GENERAL, CUSPARSE_FILL_MODE_UPPER, CUSPARSE_DIAG_TYPE_NON_UNIT, index)
+            m,n = A.dims
+            if m != n
+                throw(DimensionMismatch("A must be square, but has dimensions ($m,$n)!"))
+            end
+            mb = div(m,A.blockDim)
+            info = bsrilu02Info_t[0]
+            cusparseCreateBsrilu02Info(info)
+            @workspace size=@argout(
+                    $bname(handle(), A.dir, mb, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, A.blockDim, info[1],
+                           out(Ref{Cint}(1)))
+                )[] buffer->begin
+                    $aname(handle(), A.dir, mb, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, A.blockDim, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                    posit = Ref{Cint}(1)
+                    cusparseXbsrilu02_zeroPivot(handle(), info[1], posit)
+                    if posit[] >= 0
+                        error("Structural/numerical zero in A at ($(posit[]),$(posit[])))")
+                    end
+                    $sname(handle(), A.dir, mb, A.nnz, desc,
+                           A.nzVal, A.rowPtr, A.colVal, A.blockDim, info[1],
+                           CUSPARSE_SOLVE_POLICY_USE_LEVEL, buffer)
+                end
+            cusparseDestroyBsrilu02Info(info[1])
+            A
+        end
+    end
+end
+
+for elty in (:Float32, :Float64, :ComplexF32, :ComplexF64)
+    @eval begin
+        function ilu02(A::CuSparseMatrix{$elty},
+                       index::SparseChar)
+            ilu02!(copy(A),index)
+        end
+        function ic02(A::CuSparseMatrix{$elty},
+                      index::SparseChar)
+            ic02!(copy(A),index)
+        end
+        function ilu02(A::HermOrSym{$elty,CuSparseMatrix{$elty}},
+                       index::SparseChar)
+            ilu02!(copy(A.data),index)
+        end
+        function ic02(A::HermOrSym{$elty,CuSparseMatrix{$elty}},
+                      index::SparseChar)
+            ic02!(copy(A.data),index)
+        end
+    end
+end
